Adding in a tabs() and activate() commands to chromeless and the api documentation.
This commit exposes the target list of CDP to users of chromless via the tabs() command. That command will return
a promise which eventually resolves to an array of objects that CDP.List() returns.

Callers can then use that object to grab the targetId they want, and then activate the tab with activate(targetId).

This is my first time working with JS promises and async code, so a glance at that would be appreciated. In addition,
I imagine the naming of tabs vs targets is potentially a mine field. I have chosen tabs because that is my use case,
but I am open to changing it if others disagree.